refactor: Remove console.log from useDisableRightClickOnTouchDevice test & convert to TypeScript
