--- frontend/src/hooks/useDisableRightClickOnTouchDevices.js
@@ -1,20 +1,0 @@
-import { useEffect } from 'react';
-
-const useDisableRightClickOnTouchDevices = () => useEffect(() => {
-  const isTouchDevice = () => !!('ontouchstart' in window || !!(navigator.maxTouchPoints));
-
-  const handleContextMenu = (event) => {
-    if (isTouchDevice()) {
-      event.preventDefault();
-    }
-  };
-
-  document.addEventListener('contextmenu', handleContextMenu);
-
-  return () => {
-    document.removeEventListener('contextmenu', handleContextMenu);
-  };
-}, []);
-
-
-export default useDisableRightClickOnTouchDevices;
